Refactor: Split Dashboard.tsx into smaller components
Split the Dashboard.tsx component into PortfoliosTab, IntegrationsTab, and AccountTab components for better organization and maintainability. Removed unused imports.
