+++ src/components/dashboard/IntegrationsTab.tsx
@@ -1,0 +1,132 @@
+
+import IntegrationCard from '@/components/dashboard/IntegrationCard';
+import { useToast } from '@/components/ui/use-toast';
+import { useGitHubIntegration } from '@/hooks/useGitHubIntegration';
+import { useLinkedInIntegration } from '@/hooks/useLinkedInIntegration';
+
+export default function IntegrationsTab() {
+  const { toast } = useToast();
+  
+  const {
+    isConnected: githubConnected,
+    profile: githubProfile,
+    connectGitHub,
+    disconnectGitHub,
+    syncData: syncGitHub,
+    loading: githubLoading
+  } = useGitHubIntegration();
+
+  const {
+    isConnected: linkedinConnected,
+    profile: linkedinProfile,
+    connectLinkedIn,
+    disconnectLinkedIn,
+    syncData: syncLinkedIn,
+    loading: linkedinLoading
+  } = useLinkedInIntegration();
+
+  const handleConnectGithub = async () => {
+    try {
+      await connectGitHub();
+      toast({
+        title: "GitHub OAuth initiated",
+        description: "You'll be redirected to GitHub to authorize the connection.",
+      });
+    } catch (error) {
+      toast({
+        variant: "destructive",
+        title: "GitHub connection failed",
+        description: "There was an error initiating the GitHub OAuth flow.",
+      });
+    }
+  };
+
+  const handleDisconnectGithub = async () => {
+    try {
+      await disconnectGitHub();
+      toast({
+        title: "GitHub disconnected",
+        description: "Your GitHub account has been unlinked.",
+      });
+    } catch (error) {
+      toast({
+        variant: "destructive",
+        title: "Disconnection failed",
+        description: "There was an error disconnecting your GitHub account.",
+      });
+    }
+  };
+
+  const handleConnectLinkedin = async () => {
+    try {
+      await connectLinkedIn();
+      toast({
+        title: "LinkedIn OAuth initiated",
+        description: "You'll be redirected to LinkedIn to authorize the connection.",
+      });
+    } catch (error) {
+      toast({
+        variant: "destructive",
+        title: "LinkedIn connection failed",
+        description: "There was an error initiating the LinkedIn OAuth flow.",
+      });
+    }
+  };
+
+  const handleDisconnectLinkedin = async () => {
+    try {
+      await disconnectLinkedIn();
+      toast({
+        title: "LinkedIn disconnected",
+        description: "Your LinkedIn account has been unlinked.",
+      });
+    } catch (error) {
+      toast({
+        variant: "destructive",
+        title: "Disconnection failed",
+        description: "There was an error disconnecting your LinkedIn account.",
+      });
+    }
+  };
+
+  return (
+    <div className="mt-6">
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        <IntegrationCard 
+          name="GitHub"
+          icon={
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="h-full w-full">
+              <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4"></path>
+              <path d="M9 18c-4.51 2-5-2-7-2"></path>
+            </svg>
+          }
+          description="Sync your repositories, contributions, and coding languages."
+          isConnected={githubConnected}
+          onConnect={handleConnectGithub}
+          onDisconnect={handleDisconnectGithub}
+          loading={githubLoading}
+          lastSynced={githubProfile?.last_synced_at}
+          profileData={githubProfile}
+        />
+        
+        <IntegrationCard 
+          name="LinkedIn"
+          icon={
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="h-full w-full">
+              <path d="M16 8a6 6 0 0 1 6 6v7h-4v-7a2 2 0 0 0-2-2 2 2 0 0 0-2 2v7h-4v-7a6 6 0 0 1 6-6z"></path>
+              <rect width="4" height="12" x="2" y="9"></rect>
+              <circle cx="4" cy="4" r="2"></circle>
+            </svg>
+          }
+          description="Import your professional experience, education, and skills."
+          isConnected={linkedinConnected}
+          onConnect={handleConnectLinkedin}
+          onDisconnect={handleDisconnectLinkedin}
+          loading={linkedinLoading}
+          lastSynced={linkedinProfile?.last_synced_at}
+          profileData={linkedinProfile}
+        />
+      </div>
+    </div>
+  );
+}
